Fix locales containing a dot.
By using the negative result of `strrpos()`, the dot (and charset) where not removed in my case (locale set to `de_DE.utf-8`/`en_US.utf-8`). `strrpos()` returns the position of the dot from the start of the locale string, so as far as I understand, the correct way to call `substr()` would be `substr( $locale, 0, $x );`. At least, that resolved the issue in my case.

https://www.php.net/manual/en/function.substr.php